pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dida-cli"
version = "0.2.2"
description = "事咚咚 - 个人待办管理的 CLI 工具"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "gudong"}]
keywords = ["todo", "cli", "task-management"]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "dong-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
dida = "dida.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/dida"]
